+++ BE_TKDecor/Core/Dtos/ProductReport/ProductReportUpdateDto.cs
@@ -5,7 +5,7 @@ namespace BE_TKDecor.Core.Dtos.ProductReport
 {
     public class ProductReportUpdateDto
     { 
-        public int ProductReportId { get; set; }
+        public long ProductReportId { get; set; }
 
         [RegularExpression($"^({ReportStatusContent.Accept}|{ReportStatusContent.Reject})$")]
         public string ReportStatus { get; set;} = null!;
